Introduce Atoms (#10)
This change introduces a new Atom type that holds primitives.
There is now a default atom method on the sinks that receive
these atoms rather than individual methods for each primitive type.
This also changes the Event type so that atoms are directly placed
on it. Additionally instead of passing event references, they are
now passed by value.